Revert "add comments for hw1"
