+++ src/test/resources/testing.xml
@@ -1,0 +1,14 @@
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
+
+<suite name="TestNg Suite" parallel="methods" thread-count="3" >
+
+    <test name="Test 1" >
+
+        <classes>
+            <class name="lesson2.SimpleTestWithAnnotations2" />
+
+        </classes>
+
+    </test>
+
+</suite>
