--- test/java/lesson2/SimpleTestParallel.java
@@ -1,47 +1,0 @@
-package lesson2;
-
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.testng.annotations.BeforeClass;
-import org.testng.annotations.Test;
-
-import static java.lang.System.setProperty;
-import static org.testng.Assert.assertEquals;
-
-
-public class SimpleTestParallel {
-
-    @BeforeClass
-    public void beforeClass(){
-        setProperty("webdriver.chrome.driver","src\\main\\resources\\chromedriver.exe");
-    }
-
-    @Test
-    public void simpleTest(){
-        WebDriver driver=new ChromeDriver();
-
-        //1
-        driver.manage().window().maximize();
-
-        //2
-        driver.navigate().to("https://epam.github.io/JDI/index.html");
-
-        //3
-        assertEquals(driver.getTitle(),"Home Page");
-
-        //4
-        driver.findElement(By.cssSelector("[id='user-icon']")).click();
-        driver.findElement(By.cssSelector("[id='name']")).sendKeys("epam");
-        driver.findElement(By.cssSelector("[id='password']")).sendKeys("1234");
-        driver.findElement(By.cssSelector("[id='login-button']")).click();
-
-        //5
-        driver.findElement(By.cssSelector("[id='user-name']")).equals("PITER CHAILOVSKII");
-
-        //6
-        assertEquals(driver.getTitle(),"Home Page");
-
-        // driver.close();
-    }
-}
